% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmers.R
\name{cluster_kmers}
\alias{cluster_kmers}
\title{Cluster kmers into similar repetitive sequences}
\usage{
cluster_kmers(core_cutoff, kmer_counts)
}
\arguments{
\item{core_cutoff}{any integer value greater than 5}

\item{kmer_counts}{counts table from count_kmers()}
}
\value{
list of kmer clusters (P-cloud object)
}
\description{
Cluster kmers into similar repetitive sequences
}
\examples{
cluster_kmers(10, ecoliCounts)
}
