% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{count_kmers}
\alias{count_kmers}
\title{Count non-canonical k-mers in a FASTA file}
\usage{
count_kmers(seq_path, k = NA)
}
\arguments{
\item{seq_path}{path to FASTA file. Cannot handle N/ambiguous sequences.}

\item{k}{(OPTIONAL) length of k-mers.
Default calculated with round(log(seq_length, base = 4)) + 1.
Must be greater than default_k and less than 500.}
}
\value{
data.table of k-mers seen more than once in given FASTA file
}
\description{
Count non-canonical k-mers in a FASTA file
}
\examples{
seq_path <- fs::path_package("extdata", "phiX.fasta.gz", package = "kmer2tns")
count_kmers(seq_path = seq_path)
}
