% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers_positions.R
\name{count_kmers_positions}
\alias{count_kmers_positions}
\title{Count kmers with position information}
\usage{
count_kmers_positions(seq_path, k = NA)
}
\arguments{
\item{seq_path}{path to fasta file}

\item{k}{length of k-mers}
}
\value{
data.table with three columns: kmers, count, and positions
}
\description{
Count kmers with position information
}
\examples{
seq_path <- fs::path_package("extdata", "phiX.fasta.gz", package = "kmer2tns")
count_kmers_positions(seq_path = seq_path)
}
