% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoliCounts}
\alias{ecoliCounts}
\title{Ecoli k-mer counts data}
\format{
A data frame with 999708 rows and 2 columns:
\describe{
  \item{kmer}{12-mer sequence}
  \item{count}{number of times seen in the sequence}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/nuccore/AP022815.1>
}
\usage{
ecoliCounts
}
\description{
Counts of k-mers appearing more than once in the Ecoli genome
}
\keyword{datasets}
