% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmers.R
\name{identify_top_kmer}
\alias{identify_top_kmer}
\title{Identify top kmer in a kmer_count table}
\usage{
identify_top_kmer(table)
}
\arguments{
\item{table}{kmer_count table}
}
\value{
kmer with highest count and its count
}
\description{
Identify top kmer in a kmer_count table
}
\examples{
identify_top_kmer(ecoliCounts)
}
