% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\title{Log Messages with Timestamps}
\usage{
log_message(...)
}
\arguments{
\item{...}{Anything to be concatenated into the log message}
}
\value{
None
}
\description{
This function logs messages with a timestamp.
}
