% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmers.R
\name{new_expand_core_kmers}
\alias{new_expand_core_kmers}
\title{Expand core kmers}
\usage{
new_expand_core_kmers(table, core_list, cutoff_list)
}
\arguments{
\item{table}{kmer_counts used to generate core_table list}

\item{core_list}{list of core_tables generated}

\item{cutoff_list}{cutoffs generated automatically from core_cutoff}
}
\value{
list of expanded p_clouds
}
\description{
Expand core kmers
}
