% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_kmers.R
\name{new_identify_core_kmers}
\alias{new_identify_core_kmers}
\title{Identify core kmers from a kmer_count table}
\usage{
new_identify_core_kmers(table, max_diff)
}
\arguments{
\item{table}{kmer_count table}

\item{max_diff}{max allowed differences between kmers}
}
\value{
kmer_count table of just the most repetitive kmer and its closely related sequences
}
\description{
Identify core kmers from a kmer_count table
}
\examples{
new_identify_core_kmers(ecoliCounts, 3)
}
