% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PhiXcounts}
\alias{PhiXcounts}
\title{PhiX k-mer counts data}
\format{
A data frame with 401 rows and 2 columns:
\describe{
  \item{kmer}{8-mer sequence}
  \item{count}{number of times seen in the sequence}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/nuccore/9626372>
}
\usage{
PhiXcounts
}
\description{
Counts of k-mers appearing more than once in the PhiX genome
}
\keyword{datasets}
