% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_kmer_count_table}
\alias{validate_kmer_count_table}
\title{Validate kmer_count_table format}
\usage{
validate_kmer_count_table(kmer_counts)
}
\arguments{
\item{kmer_counts}{data.table with two columns: "kmer" and "count".}
}
\value{
TRUE if kmer_count is in the correct format. FALSE if not.
}
\description{
Validate kmer_count_table format
}
\examples{
validate_kmer_count_table(ecoliCounts)
}
